requirements.txt
altgraph==0.17.4
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
bs4==0.0.2
camelot-py==0.9.0
certifi==2024.2.2
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
et-xmlfile==1.1.0
Flask==3.0.3
frozendict==2.4.4
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
macholib==1.16.3
MarkupSafe==2.1.5
multitasking==0.0.11
mutagen==1.47.0
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.2
peewee==3.17.6
pillow==10.3.0
platformdirs==4.2.2
playwright==1.43.0
pyasn1==0.6.0
pycparser==2.22
pycryptodomex==3.20.0
pyee==11.1.0
pygame==2.5.2
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.5
PyPDF2==2.12.1
pypdfium2==4.30.0
PySimpleGUI==5.0.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
rsa==4.9
selenium==4.21.0
setuptools==69.5.1
simpleaudio==1.0.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.32
termcolor==2.5.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.3
wsproto==1.2.0
yfinance==0.2.41
yt-dlp==2024.4.9
